# swagger/openapi.yml
openapi: 3.0.3
info:
  title: Swagger Todo - OpenAPI 3.0
  version: 1.0.0
servers:
  - url: http://127.0.0.1:3000/v1
tags:
  - name: Todo-Item
    description: methods for Todo Items
  - name: User
    description: methods for Users
paths:
  /todo:
    post:
      tags:
        - Todo-Item
      summary: Add a new todo-item
      description: Add a new todo-item
      operationId: createTodoItem
      x-eov-operation-handler: todoItem-controller
      requestBody:
        description: Create a new todo-item
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoItemData'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItem'
        '405':
          description: Invalid input
  /todo/{todoItemId}:
    get:
      tags:
        - Todo-Item
      summary: Find todo-item by ID
      description: Returns a single todo-item
      operationId: getTodoItemById
      x-eov-operation-handler: todoItem-controller
      parameters:
        - name: todoItemId
          in: path
          description: ID of todo-item to return
          required: true
          schema:
            type: string
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItem'
        '400':
          description: Invalid ID supplied
        '404':
          description: Todo Item not found
    delete:
      tags:
        - Todo-Item
      summary: Deletes a Todo Item
      description: delete a Todo Item
      operationId: deleteTodoItemById
      x-eov-operation-handler: todoItem-controller
      parameters:
        - name: todoItemId
          in: path
          description: Todo Item id to delete
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Todo Item deleted
        '400':
          description: Invalid Todo value
    put:
      tags:
        - Todo-Item
      summary: Update an existing todo-item
      description: Update an existing todo-item by Id
      operationId: updateTodoItem
      x-eov-operation-handler: todoItem-controller
      parameters:
        - name: todoItemId
          in: path
          description: Todo Item id to update
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent todo-item body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoItemData'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoItem'
        '400':
          description: Invalid ID supplied
        '404':
          description: Todo-Item not found
        '405':
          description: Validation exception

  /user/{userId}/todo-items:
    get:
      tags:
        - user
      summary: List todo items for a single user
      description: ''
      operationId: getTodoItemsPerUser
      x-eov-operation-handler: todoItem-controller
      parameters:
        - name: userId
          in: path
          description: The user id for which todo items are to be fetched
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/TodoItem'
        '400':
          description: Invalid username/password supplied
components:
  schemas:
    TodoItem:
      type: object
      properties:
        todoItemId:
          type: string
        userId:
          type: string
          format: uuid
        description:
          type: string
        title: 
          type: string
        completed:
          type: boolean
    TodoItemData:
      type: object
      properties:
        userId:
          type: string
        description:
          type: string
        title: 
          type: string
        completed:
          type: boolean
    UpdateTodoItemData:
      type: object
      properties:
        description:
          type: string
        title: 
          type: string
        completed:
          type: boolean
    User:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
          example: theUser
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: '##default'